webapp/ml-service/requirements.txt
# FastAPI ML Service Requirements
# Python 3.10+

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning
tensorflow>=2.13.0
numpy>=1.24.3
pillow>=10.1.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
